Skip adding the entry if the key is "password"  (ref commit: 05353cffb746c84e2fa593d801104189605bf5d6
Enhance the security of the website by preventing the unintentional exposure of sensitive information stored in the browser's localStorage. Previously, the application would load and list all entries from the localStorage, including entries with keys like "password." This behavior posed a security risk, allowing users to manually add a "password" key to the localStorage, and subsequently, deleting it would disrupt the site's functionality.

This commit introduces a crucial improvement. Now, when loading localStorage entries, any entry with the key "password" is explicitly skipped, mitigating the risk of accidental or intentional deletion of this sensitive information. The revised logic ensures that only valid entries, excluding "password," are processed and displayed. This change contributes to a more secure handling of localStorage data and prevents potential disruptions caused by inadvertent key deletions.